msm: autosmp: import driver from dorimanx with neobuddy patches
Updates from neobuddy applied using diff and patch.
Updated 'autosmp.c' header style from MSM hotplug, which is less weird.
There might be other style parts to be adapted as well, but this is it
for now.
Credits to both developers, authored as dorimanx because it's where the
source came from.
Imported neobuddy patches from
https://github.com/neobuddy89/shamu/commits/cm-14.1-autosmp.
Restored DEFAULT_BOOST_LOCK_DUR to 500...

This commit was put together by Ícaro Hoff. I think this might be the way
to go. It is more conservative than MPDecision, but supports similar
functions that can be controlled in the userspace. I just made a few changes
seen below:

* Removed state notifier support
* slightly lowered the boost frequency from 1.728ghz to 1.4976ghz

Change-Id: I8c72f961b437a3d759c3c4a27f6f93f4f678773d
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>
Signed-off-by: RJ Trujillo <certifiedblyndguy@gmail.com>