simple thermal: add VADC/TSENS polling mode switcher
XO_THERM_PU2 is what thermal-engine wants for MSM8974 devices, although
TSENS is a great option, it's way too much erratic and forces the driver
to act a lot when it is not even needed.
Now, this simple (poor code) switcher can allow boths modes in the same
driver at user will to choose one over the other.

Change-Id: Ida53e1333c53e96ca4eed10f636571cb1120f81d
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>