ARM: dts: 15055:  Fix overexposure when capturing photos with flash
Full credits to Sultanxda <sultanxda@gmail.com>. This is just an
adaptation of his work for the multiple HW revisions in our DTS. This
includes the later fix that slightly lowers his initial voltage values
for the LED as well. Here is the original commit message he left:

"When photos are captured with flash, there are 2 phases of the snapshot:
pre-flash (low power) and the actual flash (high power). During the low-power
flash phase, 3A acquires data on the lighting and uses it to configure the
actual snapshot with high-power flash. Since the gap between low-power and
high-power flash is so great, 3A overcompensates exposure by assuming the scene
with high-power flash is much darker than it actually is. This results in the
final photo with high-power flash being overexposed and susceptible to
motion blur due to the use of a longer-than-necessary exposure time in
most cases.

"By increasing the current of the LED for torch-mode from 127 mA to [190] mA,
the pre-flash phase when capturing a photo appears brighter to 3A, so 3A
won't overestimate the exposure. Photos with flash are now properly exposed.

"Additionally, the torch is now [~1.5x] brighter."

Change-Id: I2b942ff8aecd8686fea89a69e4a31e74aca161fc