Update patch set 2
Patch Set 2:

I'd just like to mention that I have gotten 4.9 sources working, however only with higher GCC revisions as fs_inode was introduced but only utilized by F2FS. So it might be possible to do if I convert it over yo the appopriate inode. Just a suggestion as the 4.4 F2FS sources haven't been updated for a long time (2015). It would however require a bit of work on my part, but I'm okay with it if we do want to go in that direction. What do you think?

Patch-set: 2