msm: thermal: import my enhanced version of Franco's driver
The default thermal implementation relies upon a thermal engine,
several blobs, and at times, is not very effective at all. It is
not dynamic and can throttle when it is not needed. My attempt here
avoids that from happening.

*There are 12 steps. These match the values Sultan uses in his driver
*It can offline CPU cores. By default, this occurs first at 46 degrees
*We can make it even more dynamic by using sysfs nodes in the future.
 I chose not to at the moment, because I assume several people use Kernel
 Auditor where the values do not properly appear
*This uses VADC polling which is more effective, and keeps the device
 cooler.

I started work on this months ago, and it has become more advanced since.

I think it is a perfect candidate to replace the default implementation
with this one as it works far better, and is a much better solution than
what is shipped on Hammerhead currently with my adjustments

Change-Id: I2b436cda92b28c4ccdb42eeb3fc342c58c54bab8