Update patch set 1
Patch Set 1: Code-Review-1

The Sony developers I've talked to say this is a regression compared to mpdecision, and that the only advantage is the fact that it's open source. If that is the case, I don't want this merged. Open source is only an advantage if it is greater than or equal to the proprietary option.

Patch-set: 1
Reviewer: Gerrit User 1000012 <1000012@3162e3d8-d894-470e-abd9-724f7b0effad>
Label: Code-Review=-1