Update patch set 1
Patch Set 1:

I am unhappy with MPDecision because it seems too sporadic when it isn't necessary. CPUQuiet accomplishes the same task conservatively with much less overhead, and drain on the battery. CPU boost accomplishes what the boost function of MPDecision is there for, and CPUQuiet can even make use of that function through RQ stats if it is wanted. Those are my thoughts though. The driver does remain disabled which is a good thing to keep in mind too unless activated in the userspace via sysfs

Patch-set: 1