Update patch set 1
Patch Set 1:

Here is the issue with that Jake. I tried your volume boost configuration to address this and still got complaints. We can disable fluence though as I proposed earlier. We don't need it with the CAF audio HAL anyway. That should give a boost to audio. This is still useful though. Completely up to you guys however. I can always fork the source and do what Icaro did for the G2 with a more minimal aim. But are you okay with disabling fluence at the very least?

Patch-set: 1