Update patch set 1
Patch Set 1: Code-Review-1

Wasn't the purpose in this to get rid of the 1ms delay in updates? If we add that back, then the delay would still partially be there. I don't think there is a point in doing this. Latency might actually be lower the way it is

Patch-set: 1