Update patch set 1
Patch Set 1: Code-Review-1

The point is that this change does more than it says it does. If you want to change the value from 1 to 0 that belongs to a seperate change.

About this change, I am not sure whether it's a good idea as the critical kgsl parts are already moved to a SCHED_FIFO worker and we don't want to have too many workqueues as high priority if not necessary.

Patch-set: 1
Reviewer: Gerrit User 1000011 <1000011@3162e3d8-d894-470e-abd9-724f7b0effad>
Label: Code-Review=-1