Update patch set 1
Patch Set 1:

Those are two different drivers. I have that change. It is specific to Bacon however. The s3508 works a lot differently. I have no idea why OnePlus even put the TS configuration for Bacon in synaptics_dsx to begin with. That was back when Onyx and Bacon shared the same kernel. Sorry for the confusion

Patch-set: 1