Update patch set 1
Patch Set 1:

I'll give it a look over again. The issue arises from I've already invested my time in a lot of other changes that really help with latency overall, and I'd prefer to keep it that way rather than ditching hours of work to port something over that does the same thing. Its sentimental, I get that, but if a regression is caused, I see no reason to

Patch-set: 1
Label: Code-Review=0