Update patch set 1
Patch Set 1:

Some users are reporting that onyx will not charge to 100% on recent builds and instead gets stuck at 99%.  I suspect that this commit is the reason, as it is the only recent power commit.  If this is not it, it is probably something else from the July 17th commits.  In the future, please have Alex or I approve commits before merging, especially large amounts such as that.

Patch-set: 1
Reviewer: Gerrit User 1000012 <1000012@3162e3d8-d894-470e-abd9-724f7b0effad>
Label: Code-Review=0