kernel: Do not verify the persistance of the su daemon
This is pointless on devices that do not comply with CTS standards already,
and by limiting the access of the su daemon, several other holes are open to
expose it.

Not only that, but if this were an actual issue, it would be fixed in mainline
Linux due to security concerns.

Change-Id: Idbc504cd75c6dd8ec14706dfd262385aa586e58b